Prepared to move "too large" functionality.
At the end of finishing the previous commit, I thought that it might be
better to move the "too large" value check to `OhmValueCalculator` since
the `IOhmValueCalculator` specifies return an `int`.

This partitioning will allow the `Resistor` class to return 64-bit
values for `Resistance` and letting callers decide what to do with these
large values.

This commit prepares for that movment by add a "too large" value test
for `OhmValueCalculator`.