(BROKEN) Correct handling of invalid bands
I am working on handling of invalid bands. Some of these bands are
invalid because no value is supplied (for example, a Gold significant
figure band A). Other bands are invalid because I have chosen not to
handle the decimal tolerances (gold and silver).

Although the code compiles, I have chosen to check it in so that I can
refactor code before moving on to the additional broken unit tests.

This commit refactors our validation that significant figure bands are
present.