Add calculation feature
This commit adds the ability to calculate the resistance of a resistor.
It uses a "post-back" approach by posting the values back to the server.
In future iterations, we can change this to use AJAX to avoid these
trips to the server.

In addition, this increment uncovered errors:

* Given I have chosen
	| Band C |
	| Gray   |
	| White  |
	| Gold   |
	| Silver |
* When I press the "Calculate" button
* Then the application raises an exception

This behavior is not unexpected. Although these four colors are valid
(see https://en.wikipedia.org/wiki/Electronic_color_code), they may
generate resistances that are non-integral. (For example, Gray and White
may generate long values and Gold and Silver may generated values with
decimals.) When I noticed this issue in my domain model testing, I chose
to have the `Resistor` class return long values, throwing an exception
if we could generate a decimal value. The implementation of
`IOhmValueCalculator` implements the specified interface and throw
exceptions otherwise. I will need to address this issue in future
increments.