Change form to use AJAX to update resistance
Previous commits used `Html.BeginForm` to submit a "POST" request to an
action to change the calculated resistance on pressing the "Calculate"
button.

This commit changes the application to use AJAX to perform a get request
that replaces only the resistance value on the page.

This change introduced an irritating visual defect. When one first
presses the "Calculate" button, the application resizes the label.
Subsequent presses **do not** change the size of the text. This same
behavior occurs in both Google Chrome and Internet Explorer so its most
likely with my code.

Even though this issue is incredibly visually jarring, because of time,
I choose to address other "low-hanging fruit" like the About and Contact
pages.